\name{ExtChiarella-package}
\alias{ExtChiarella-package}
\alias{ExtChiarella}
\docType{package}
\title{
\packageTitle{ExtChiarella}
}
\description{
\packageDescription{ExtChiarella}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ExtChiarella}
\packageIndices{ExtChiarella}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ExtChiarella}

Maintainer: \packageMaintainer{ExtChiarella}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
