% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaSim.R
\name{ExtendedNLChiarella}
\alias{ExtendedNLChiarella}
\title{Title}
\usage{
ExtendedNLChiarella(
  N,
  kappa,
  kappa_3,
  beta,
  sigma_n,
  sigma_v,
  g,
  v0,
  p0,
  m0,
  alpha,
  gamma
)
}
\arguments{
\item{N}{Length of trajectory}

\item{kappa}{Linear coeff for fundamentalist demand function}

\item{kappa_3}{cubic term coefficient for fundamentalist demand function}

\item{beta}{Chartist Coefficient}

\item{sigma_n}{Noise trader volitility}

\item{sigma_v}{Value signal volitility}

\item{g}{Growth}

\item{v0}{Initial fundamental value}

\item{p0}{Initial Price}

\item{m0}{Initial momentum}

\item{alpha}{Smoothing parameter}

\item{gamma}{Trend signal saturation parameter}
}
\value{

}
\description{
Title
}
\examples{
f
}
