% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaSim.R
\name{ExtendedvolChiarella}
\alias{ExtendedvolChiarella}
\title{Title}
\usage{
ExtendedvolChiarella(
  N,
  kappa,
  kappa_3,
  rho,
  beta,
  sigma_n,
  sigma_v,
  g,
  v0,
  p0,
  m0,
  S0,
  voltar,
  alpha_1,
  alpha_2,
  gamma
)
}
\arguments{
\item{N}{desc}

\item{kappa}{desc}

\item{kappa_3}{desc}

\item{rho}{desc}

\item{beta}{desc}

\item{sigma_n}{v}

\item{sigma_v}{desc}

\item{g}{desc}

\item{v0}{desc}

\item{p0}{desc}

\item{m0}{desc}

\item{S0}{desc}

\item{voltar}{desc}

\item{alpha_1}{desc}

\item{alpha_2}{desc}

\item{gamma}{desc}
}
\value{

}
\description{
Title
}
\examples{
f
}
