% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{grad_EC}
\alias{grad_EC}
\title{Title}
\usage{
grad_EC(kappa, beta, sigma_N, p1, v1, v0, t0)
}
\arguments{
\item{kappa}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{p1}{description}

\item{v1}{description}

\item{v0}{description}

\item{t0}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
