% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{grad_NLEC}
\alias{grad_NLEC}
\title{Title}
\usage{
grad_NLEC(kappa, kappa3, beta, sigma_N, sigma_v, g, p1, v1, v0, t0)
}
\arguments{
\item{kappa}{description}

\item{kappa3}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{sigma_v}{description}

\item{g}{description}

\item{p1}{description}

\item{v1}{description}

\item{v0}{description}

\item{t0}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
