% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{grad_volNLEC}
\alias{grad_volNLEC}
\title{Title}
\usage{
grad_volNLEC(
  kappa,
  kappa3,
  beta,
  sigma_N,
  sigma_v,
  g,
  tau,
  rho,
  p1,
  v1,
  v0,
  t0,
  S,
  u
)
}
\arguments{
\item{kappa}{description}

\item{kappa3}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{sigma_v}{description}

\item{g}{description}

\item{tau}{description}

\item{rho}{description}

\item{p1}{description}

\item{v1}{description}

\item{v0}{description}

\item{t0}{description}

\item{S}{description}

\item{u}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
