% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaSim.R
\name{Kalmanfs}
\alias{Kalmanfs}
\title{Title}
\usage{
Kalmanfs(p, kappa, beta, gamma, sigma_n, sigma_v, sigma_0, g, alpha, v0, t0)
}
\arguments{
\item{p}{price signal}

\item{kappa}{Fundamentalist demand coefficient}

\item{beta}{Chartist coefficient}

\item{gamma}{Saturation parameter}

\item{sigma_n}{Noise trader volitility}

\item{sigma_v}{Value signal volitility}

\item{sigma_0}{initial variance}

\item{g}{Growth}

\item{alpha}{smoothing parameter}

\item{v0}{Initial fundamental value}

\item{t0}{Burn in time}
}
\value{

}
\description{
Title
}
\examples{
f
}
