% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{LaplaceNLApprox}
\alias{LaplaceNLApprox}
\title{Title}
\usage{
LaplaceNLApprox(
  kappa,
  kappa3,
  beta,
  sigma_N,
  sigma_v,
  g,
  p1,
  epsilon,
  vinit,
  v0,
  t0
)
}
\arguments{
\item{kappa}{description}

\item{kappa3}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{sigma_v}{description}

\item{g}{description}

\item{p1}{description}

\item{epsilon}{description}

\item{vinit}{description}

\item{v0}{description}

\item{t0}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
