% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{LaplacevolNLApprox}
\alias{LaplacevolNLApprox}
\title{Title}
\usage{
LaplacevolNLApprox(
  kappa,
  kappa3,
  beta,
  sigma_N,
  sigma_v,
  tau,
  rho,
  g,
  p1,
  epsilon,
  vinit,
  v0,
  t0,
  alpha1,
  alpha2
)
}
\arguments{
\item{kappa}{description}

\item{kappa3}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{sigma_v}{description}

\item{tau}{description}

\item{rho}{description}

\item{g}{description}

\item{p1}{description}

\item{epsilon}{description}

\item{vinit}{description}

\item{v0}{description}

\item{t0}{description}

\item{alpha1}{description}

\item{alpha2}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
