% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{LinearNewton}
\alias{LinearNewton}
\title{Title}
\usage{
LinearNewton(kappa, beta, sigma_N, p, epsilon, vinit, v0, t0)
}
\arguments{
\item{kappa}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{p}{description}

\item{epsilon}{description}

\item{vinit}{description}

\item{v0}{description}

\item{t0}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
