% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaSim.R
\name{msignal}
\alias{msignal}
\title{Title}
\usage{
msignal(alpha, p, t0)
}
\arguments{
\item{alpha}{Smoothing parameter}

\item{p}{price}

\item{t0}{smoothing start time}
}
\value{

}
\description{
Title
}
\examples{
f
}
