% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R,
%   R/ExtendedChiarellaSim.R
\name{NL}
\alias{NL}
\title{Title}
\usage{
NL(kappa, kappa_3, x)

NL(kappa, kappa_3, x)
}
\arguments{
\item{kappa}{Linear coeff}

\item{kappa_3}{Cubic term coeff}

\item{x}{Argument}
}
\value{



}
\description{
Title

Title
}
\examples{
f
f
}
