% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaFitting.R
\name{NonLinear_volNewton}
\alias{NonLinear_volNewton}
\title{Title}
\usage{
NonLinear_volNewton(
  kappa,
  kappa3,
  beta,
  sigma_N,
  sigma_v,
  g,
  tau,
  rho,
  p,
  epsilon,
  vinit,
  v0,
  t0,
  S,
  u
)
}
\arguments{
\item{kappa}{description}

\item{kappa3}{description}

\item{beta}{description}

\item{sigma_N}{description}

\item{sigma_v}{description}

\item{g}{description}

\item{tau}{description}

\item{rho}{description}

\item{p}{description}

\item{epsilon}{description}

\item{vinit}{description}

\item{v0}{description}

\item{t0}{description}

\item{S}{description}

\item{u}{description}
}
\value{

}
\description{
Title
}
\examples{
f
}
