% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtendedChiarellaSim.R
\name{volsig}
\alias{volsig}
\title{Title}
\usage{
volsig(p, t0, alpha_2)
}
\arguments{
\item{p}{log price}

\item{t0}{smoothing start time}

\item{alpha_2}{smoothing parameter}
}
\value{

}
\description{
Title
}
\examples{
f
}
